fix(CIS) use `awk` instead of `grep`
Using `grep -v <pattern>` will return a non-zero code when
it matches the pattern. This non-zero return was causing
the command group to fail and Ansible marking this task as failing.

The reason we are masking out the `/var/lib/docker` path is because
we don't want the `setuid`/`setgid` binaries in the Docker container
overlays to be monitored by `auditd`.