fix(cis) Add task to ensure Match User block ends
addresses #294169

Fix an issue breaking /etc/sshd_config if the Ubuntu 20.04 CIS
hardening rules are applied to a system where the AAD SSH Login
extension has been installed.

The changes made to /etc/sshd_config by AAD SSH Login do not
terminate a `Match User` block leading to all following sshd_config
lines to be incorrectly interpreted as in a `Match` block.

Many directives that follow are not allowed in a `Match` block.

By inserting a `Match all` block, we pop out of a `Match` context
and back into a "global" block.